Changes the misnomer "collection" to "vehicle"
Vehicles, PIDs, and Functions each have their own “collection” of
information.
I decided I did not like that “collection” was used as the name for the
group that vehicles are stored in as it is not descriptive of what is
actually being stored in that collection.
Thus, I changed all references from “collection” to “vehicle”