Implements changes to make a paradigm change
I have decided to change how the data is stored in the website.
Instead of having things stored in arrays and iterating with ng-repeat
and a range, I am going to store things in a dictionary style format. I
am using the objects ids  and ng-repeat to iterate through sub records.