Implements method to get all unique function description tags
This method is very “brute force”
The method does a find for all the function, then for each function it
iterates over each list of tags and adds the tag to an array, if the
tag is not already in that array.
I intuitively feel that there is a better way to do this, but for
respect to time I will not try to find that right now.